MAHOUT-479:  Implemented ClusterClassifier, ClusterPolicy(s) and ClusterIterator which can duplicate k-means and Dirichlet clustering in sequential execution only. Added unit tests and switched DisplayKMeans and DisplayDirichlet to use the ClusterIterator. Gives a pretty good sanity check of the clustering. Changed pdf() implementation of GaussianCluster to use the product of the component pdfs vs the average. Seems to work a lot better. Deprecated a number of old Dirichlet models and the experimental VectorModelClassifier. Changed the type of AbstractCluster numPoints to long from int All unit tests run
git-svn-id: https://svn.apache.org/repos/asf/mahout/trunk@1094222 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/mahout/clustering/Cluster.java
@@ -61,7 +61,7 @@ public interface Cluster extends Model<VectorWritable>, Parametered {
    * Get an integer denoting the number of points observed by this cluster
    * @return an integer 
    */
-  int getNumPoints();
+  long getNumPoints();
   
   /**
    * Produce a custom, human-friendly, printable representation of the Cluster.
